requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
matplotlib
Pillow
streamlit
requests
joblib
tqdm
